Updated MPL to use slam::maps::instead of envire.
Mainly replaced uses of envire::TraversabilityGrid with
slam::maps::TraversabilityGrid and made changes to accomodate
this transition.

**Work in Progress**
Requires Testing

+++ test/test_SbplSplineMotionPrimitives.cpp
@@ -46,7 +46,7 @@ BOOST_AUTO_TEST_CASE(generate_end_angles)
     for(int i = 0; i < config.numAngles; ++i)
     {
         vector<int> endAngles = prims.generateEndAngles(i, config);
-        BOOST_CHECK(endAngles.size() <= config.numEndAngles);
+        BOOST_CHECK(static_cast<int>(endAngles.size()) <= config.numEndAngles);
         
         set<int> angleSet;
         for(int angle : endAngles)
